requirements.txt
anyio==4.7.0
certifi==2024.12.14
charset-normalizer==3.4.0
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
nest-asyncio==1.6.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
python-telegram-bot==21.9
requests==2.32.3
schedule==1.2.2
sniffio==1.3.1
tenacity==9.0.0
typing_extensions==4.12.2
urllib3==2.3.0
